feat(view-stops-in-map): move route logic to reducer
