+++ src/store/selectors/stopSelectors.ts
@@ -1,8 +1,10 @@
 import { createSelector } from "reselect";
 import { StopLocation } from "../../api/trimet/types";
 import { RootState } from "../reducers";
+import { RouteDirection } from "../reducers/stopsReducer";
 
 const allStopLocations = (state: RootState) => state.stopsReducer.stopLocations;
+const allNearbyRoutes = (state: RootState) => state.stopsReducer.nearbyRoutes;
 
 const stopLocationByLocationId = (state: RootState, locationId: number) =>
   state.stopsReducer.stopLocations[locationId];
@@ -20,6 +22,11 @@ export const stopsLoadingSelector = createSelector(
 export const allStopLocationsSelector = createSelector(
   allStopLocations,
   (stopLocations: StopLocation[]) => stopLocations
+);
+
+export const allNearbyRoutesSelector = createSelector(
+  allNearbyRoutes,
+  (nearbyRoutes: RouteDirection[]) => nearbyRoutes
 );
 
 export const timeOfLastLoadSelector = createSelector(
