+++ src/view/stops/components/Stops.tsx
@@ -1,7 +1,6 @@
 import { map } from "lodash";
 import React from "react";
 import { Route, StopLocation } from "../../../api/trimet/types";
-import NearbyStopsMap from "../../../component/maps/NearbyStopsMap";
 import { StopLocationsDictionary } from "../../../store/reducers/stopsReducer";
 import StopContainer from "../containers/StopContainer";
 import "../Stops.css";
@@ -10,7 +9,6 @@ interface Props {
   stopLocations: StopLocationsDictionary;
   showArrivals: boolean;
   onRouteIndicatorClick: (route: Route) => void;
-  currentLocation: number[];
 }
 
 export default class Stops extends React.Component<Props> {
@@ -34,12 +32,7 @@ export default class Stops extends React.Component<Props> {
   }
 
   public render() {
-    const {
-      stopLocations,
-      showArrivals,
-      onRouteIndicatorClick,
-      currentLocation
-    } = this.props;
+    const { stopLocations, showArrivals, onRouteIndicatorClick } = this.props;
 
     if (!stopLocations) {
       return null;
@@ -47,10 +40,6 @@ export default class Stops extends React.Component<Props> {
 
     return (
       <div className="stops-wrapper">
-        <NearbyStopsMap
-          currentLocation={currentLocation}
-          stopLocations={stopLocations}
-        />
         {Stops.getLocationInfo(
           stopLocations,
           showArrivals,
